# storage: Implement limit in mergeGenericQuerier