<!--
    Please give your PR a title in the form "area: short description".  For example "tsdb: reduce disk usage by 95%"

    If your PR is to fix an issue, put "Fixes #issue-number" in the description.

    Don't forget!

    - Please sign CNCF's Developer Certificate of Origin and sign-off your commits by adding the -s / --signoff flag to `git commit`. See https://github.com/apps/dco for more information.

    - If the PR adds or changes a behaviour or fixes a bug of an exported API it would need a unit/e2e test.

    - Where possible use only exported APIs for tests to simplify the review and make it as close as possible to an actual library usage.

    - Performance improvements would need a benchmark test to prove it.

    - All exposed objects should have a comment.

    - All comments should start with a capital letter and end with a full stop.
 -->

### Changes:
- A small optimization to reduce the time taken to merge label names and label values when the limit hint is set.
- Merge sort is expected to faster if the strings are truncated first before the merge.
- Added a CHANGELOG entry for the limit implementation which was missed in the previous PR

Fixes: #14161

### Benchmark
```
(base) ➜  prometheus git:(implement_metadata_limit) ✗ benchstat nolimit.txt limit.txt
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/storage
                                       │    nolimit.txt    │                 limit.txt                 │
                                       │      sec/op       │      sec/op        vs base                │
MergeLabelValuesWithLimit/benchmark-10   0.00744450n ± 24%   0.00003680n ± 24%  -99.51% (p=0.000 n=10)

                                       │ nolimit.txt │           limit.txt            │
                                       │    B/op     │    B/op     vs base            │
MergeLabelValuesWithLimit/benchmark-10    0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                                       │ nolimit.txt │           limit.txt            │
                                       │  allocs/op  │ allocs/op   vs base            │
MergeLabelValuesWithLimit/benchmark-10    0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

### Related
https://github.com/prometheus/prometheus/issues/12795